+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/home/InventoryScreen.kt
@@ -80,13 +80,13 @@ fun InventoryScreen(navController: NavHostController) {
                             )
                         "supplies" -> {
                             SupplyListScreen(
+                                onBackClick = { selectedOption = null },
                                 onSupplyClick = { id ->
                                     navController.navigate("supply/$id")
                                 },
                                 onAddSupplyClick = {
                                     navController.navigate(Screen.SupplyAdd.route)
                                 },
-                                onBack = { selectedOption = null },
                             )
                         }
                     }