+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/productBatchesList/ProductBatchesListScreen.kt
@@ -108,14 +108,6 @@ fun ProductBatchesListScreen(
                         )
                     }
                 },
-                actions = {
-                    NotificationIcon(
-                        modifier =
-                            Modifier
-                                .padding(horizontal = 24.dp)
-                                .size(24.dp),
-                    )
-                },
                 colors =
                     TopAppBarDefaults.topAppBarColors(
                         containerColor = Background,