+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/user/UserListScreen.kt
@@ -103,11 +103,6 @@ fun UserListScreen(
                         TopAppBarDefaults.topAppBarColors(
                             containerColor = Color(0xFF040610),
                         ),
-                    actions = {
-                        IconButton(onClick = { /* TODO: Notifications action */ }) {
-                            NotificationIcon()
-                        }
-                    },
                 )
 
                 // Second row: Search input and Filter button