+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## [0.1.0] - 20/11/2025
 ### Added
 - US for filter and order workshops.
+- Functional go back button to Supplies List Screen
+- Functional go back buttons to Orders and Inventory Screens
+
+### Deleted
+- All the NotificationIcon buttons of the screens
+- Validations for supply add and update
+- Scroll in detail view of a supply
+- Scroll for login, home and inventory screens.
 
 ### Fixed
 - Fix beneficiary list navbar
@@ -9,6 +17,20 @@
 - Fix Workshops navbar
 - Fix beneficiary registration navBar appearing
 - Fix view after deleting a beneficiary
+- Fix date input
+- Fix scroll bar for supply batch views
+- Input Validations
+- Fix view to not show the navbar in beneficiary details
+- Fix beneficiary list navbar
+- Fix the go back buttons of Products, Product Batches and Supplies to be the same
+- Fix the go back button of teh Beneficiary Screen to be the same icon of the Workshop screen
+- Fix data load on modify product batch screen
+- Fix expiration date availabity for null values
+- Fix modify workshops for date and hour
+- Fix workshops scroll and notifications
+- Fixed Date on Register Beneficiary now using the calendar
+- Return button on Register Beneficiary
+- Errors now indicate whats wrong on Register beneficiary
 
 ## [0.1.0] - 19/11/2025
 ### Added